requirements.txt
streamlit
google-genai
python-dotenv
langchain-chroma
langchain-google-genai
langchain-openai
langgraph
litellm
pypdf
chromadb
tiktoken
langchain-community
